requirements.txt
streamlit
google-generativeai
langchain
langchain-community
duckduckgo-search
pypdf
reportlab
python-dotenv
altair<6
